fix bug num_lines conversacion

+++ modules/conversacion.h
@@ -41,7 +41,7 @@ struct Conversacion {
 		deserialStr(str);
 	}
 	// Retornar como Json
-	std::string asJson(int num_lines=0) const;
+	std::string asJson(size_t num_lines=0) const;
 	// Agregar mensaje
 	void postear(const std::string& autor, const std::string& msg);
